requirements.txt
line-bot-sdk
psycopg2
python-dateutil
requests
fastapi
uvicorn